feat: unify memory statistics API signatures (#2388)
- GetMemoryUsage: int64_t -> uint64_t (memory size is never negative)
- GetMemoryUsageDetail: std::string (JSON) -> std::unordered_map<std::string, uint64_t>
  - Returns structured component-level memory breakdown instead of
    opaque JSON string; no dependency on internal JsonType in public API
  - HGraph implementation now uses GetMemoryUsage() per component instead
    of CalcSerializeSize(), ensuring consistency with GetMemoryUsage() total
  - SINDI now returns empty map instead of empty string
- GetEstimateBuildMemory -> EstimateBuildMemory: int64_t -> uint64_t,
  dropped redundant Get prefix for consistency with EstimateMemory
- current_memory_usage_ atomic type: int64_t -> uint64_t
- Updated all index implementations, datacell layer, tests, and eval tools

+++ src/algorithm/hgraph/hgraph.cpp
@@ -670,7 +670,7 @@ HGraph::cal_memory_usage() {
     }
 
     std::unique_lock lock(this->memory_usage_mutex_);
-    this->current_memory_usage_.store(static_cast<int64_t>(memory));
+    this->current_memory_usage_.store(memory);
 }
 
 }  // namespace vsag
